# docs/swagger.yaml
definitions:
  entity.AddBillReq:
    properties:
      bill_account_number:
        type: string
      bill_attachments:
        items:
          $ref: '#/definitions/entity.Attachment'
        type: array
      bill_bank_name:
        type: string
      bill_due_date:
        type: string
      bill_items:
        items:
          $ref: '#/definitions/entity.ItemPurchase'
        type: array
      bill_notes:
        type: string
      bill_order_number:
        type: string
      bill_shipping_cost:
        type: number
      bill_start_date:
        type: string
      bill_type:
        type: string
      supplier_id:
        type: integer
    type: object
...
  entity.Attachment:
    properties:
      attachment_file:
        items:
          type: integer
        type: array
      attachment_name:
        type: string
    type: object
  entity.BillDetailsResp:
    properties:
      bill_attachments:
        items:
          $ref: '#/definitions/entity.Attachment'
        type: array
      bill_due_date:
        type: string
      bill_items:
        items:
          $ref: '#/definitions/entity.ItemBill'
        type: array
      bill_number:
        type: string
      bill_order_number:
        type: string
      bill_shipping_cost:
        type: number
      bill_start_date:
        type: string
      bill_status:
        type: string
      bill_subtotal:
        type: integer
      bill_total:
        type: integer
      bill_type:
        type: string
    type: object
...
  entity.BillUpdateStatusReq:
    properties:
      bill_status:
        type: string
    type: object
  entity.ErrRespController:
    properties:
      err_message:
        type: string
      source_function:
        type: string
    type: object
  entity.GetBalanceAmountResp:
    properties:
      net_amount:
        type: number
    type: object
  entity.ItemBill:
    properties:
      item_amount:
        type: number
      item_description:
        type: string
      item_name:
        type: string
      item_price:
        type: number
      item_qty:
        type: integer
    type: object
  entity.ItemPurchase:
    properties:
      item_discount:
        type: number
      item_id:
        type: integer
      item_qty:
        type: integer
    type: object
...
  entity.PagedResults:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total_records:
        type: integer
    type: object
  entity.StatusResponse:
    properties:
      status:
        type: string
    type: object
...
  model.VSupplierBill:
    properties:
      bill_due_date:
        type: string
      bill_id:
        type: integer
      bill_number:
        type: string
      bill_order_number:
        type: string
      bill_start_date:
        type: string
      bill_status:
        type: string
      bill_total:
        type: number
      bill_type:
        type: string
      supplier_id:
        type: integer
      supplier_name:
        type: string
      supplier_type:
        type: string
    type: object
info:
  contact: {}
  title: Test Accounting App
  version: "1.0"
paths:
...
  /balance/header:
    get:
      consumes:
      - application/json
      description: get balance in net amount (balance - paid bills)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetBalanceAmountResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrRespController'
      summary: Get Balance Header
      tags:
      - Balance
  /bill:
    post:
      consumes:
      - application/json
      parameters:
      - description: add bill request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.AddBillReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrRespController'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrRespController'
      summary: Register Bill
      tags:
      - Bill
  /bill/{billId}:
    delete:
      consumes:
      - application/json
      description: delete bill that are not paid yet
      parameters:
      - description: bill id
        in: path
        name: billId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrRespController'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrRespController'
      summary: Delete Bill
      tags:
      - Bill
    get:
      consumes:
      - application/json
      parameters:
      - description: bill id
        in: path
        name: billId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.BillDetailsResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrRespController'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrRespController'
      summary: Get Bill Details
      tags:
      - Bill
    put:
      consumes:
      - application/json
      parameters:
      - description: bill id
        in: path
        name: billId
        required: true
        type: integer
      - description: update bill status request (paid/cancelled)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.BillUpdateStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrRespController'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrRespController'
      summary: Update Bill Status
      tags:
      - Bill
...
  /bills:
    get:
      consumes:
      - application/json
      parameters:
      - description: page requested (defaults to 0)
        in: query
        name: page
        type: integer
      - description: number of records in a page  (defaults to 20)
        in: query
        name: pagesize
        type: integer
      - description: asc / desc
        in: query
        name: order
        type: string
      - description: filter by bill status
        in: query
        name: status
        type: string
      - description: filter by supplier name
        in: query
        name: vendor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entity.PagedResults'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/model.VSupplierBill'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrRespController'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrRespController'
      summary: List All Bill
      tags:
      - Bill
...
swagger: "2.0"